refactor: move player data to different file
