+++ src/lib/players.ts
@@ -1,0 +1,10 @@
+// Possible player names
+export const PLAYER_NAMES = ["red", "green", "blue", "white"];
+
+// Look-up table for rendering player colors
+export const PLAYER_COLORS: { [key: string]: { textColor: string, bgColor: string }} = {
+  "red": { textColor: "text-red-700", bgColor: "bg-red-600"},
+  "green": { textColor: "text-lime-600", bgColor: "bg-lime-400"},
+  "blue": { textColor: "text-cyan-800", bgColor: "bg-cyan-600"},
+  "white": { textColor: "text-stone-700", bgColor: "bg-stone-100"},
+};
